Fix most tests by simplifying JSON serializing/deserializing
We now use plain ((nested) lists of) GeographicPositions as the "Coordinates" member of Polygons, MultiPolygons etc. instead of full GeoJSON objects, e.g. LineStrings.

This is more accurate to the standard (e.g. a Polygon's coordinates property is a "position array" by definition) and allows to rely on Newtonsoft JSON's standard behavior in most cases.